Refactor some game logic, complete module methods
Some of the game logic (revolving around hitting vs. standing) was
implemented in the Person class originally. This has been refactored
into the Game class. As part of the refactoring, a couple additional
methods were implemented, and some of the methods called during the game
were switched to more appropriate methods.

As well, a couple bugs were fixed regarding checking for user input.